backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.52.0",
    "python-dotenv>=1.2.1",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "pytest>=7.4.0",
    "hypothesis>=6.88.0",
    "httpx>=0.25.0",
    "supabase>=2.24.0",
]
